+++ modules/elasticsearch/redirect.tf
@@ -1,45 +1,38 @@
-# Simple Job in Nomad to redirect users to the very unfriendly Kibana URL
-
-data "aws_route53_zone" "internal" {
-  count = "${var.use_redirect ? 1 : 0}"
-  name  = "${var.es_base_domain}"
-}
-
 # Define the DNS record to point to the external LB
 resource "aws_route53_record" "redirect" {
   count = "${var.use_redirect ? 1 : 0}"
 
-  zone_id = "${data.aws_route53_zone.internal.zone_id}"
-  name    = "${local.redirect_domain}"
+  zone_id = "${var.redirect_route53_zone_id}"
+  name    = "${var.redirect_domain}"
   type    = "A"
 
   alias {
-    name                   = "${var.redirect_alias_name}"
-    zone_id                = "${data.aws_route53_zone.internal.zone_id}"
+    name                   = "${var.lb_cname}"
+    zone_id                = "${var.lb_zone_id}"
     evaluate_target_health = false
   }
 }
 
-data "template_file" "redirect_jobspec" {
+resource "aws_lb_listener_rule" "redirect" {
   count = "${var.use_redirect ? 1 : 0}"
 
-  template = "${file("${path.module}/templates/redirect.nomad")}"
+  listener_arn = "${var.redirect_listener_arn}"
+  priority     = "${var.redirect_rule_priority}"
 
-  vars {
-    region     = "${var.redirect_job_region}"
-    az         = "${jsonencode(var.redirect_job_vpc_azs)}"
-    node_class = "${var.nomad_clients_node_class}"
+  action {
+    type = "redirect"
 
-    elasticsearch_service  = "${var.es_consul_service}"
-    redirect_domain        = "${local.redirect_domain}"
-    redirect_job_name      = "${var.redirect_job_name}"
-    redirect_nginx_version = "${var.redirect_nginx_version}"
+    redirect {
+      host        = "${local.endpoint}"
+      path        = "/_plugin/kibana/#{path}"
+      port        = 443
+      protocol    = "HTTPS"
+      status_code = "HTTP_301"
+    }
   }
-}
 
-resource "nomad_job" "redirect" {
-  count = "${var.use_redirect ? 1 : 0}"
-
-  depends_on = ["consul_service.es"]
-  jobspec    = "${data.template_file.redirect_jobspec.rendered}"
+  condition {
+    field  = "host-header"
+    values = ["${aws_route53_record.redirect.fqdn}"]
+  }
 }