+++ modules/elasticsearch/slow_index.tf
@@ -1,3 +1,11 @@
+resource "aws_cloudwatch_log_group" "es_slow_index_log" {
+  count = "${var.enable_slow_index_log ? 1 : 0}"
+
+  name              = "${var.slow_index_log_name}"
+  retention_in_days = "${var.slow_index_log_retention}"
+  tags              = "${merge(var.slow_index_additional_tags, map("Name", format("%s", var.slow_index_log_name)))}"
+}
+
 data "aws_iam_policy_document" "es_slow_index_log" {
   statement {
     actions = [
@@ -5,7 +13,7 @@ data "aws_iam_policy_document" "es_slow_index_log" {
       "logs:CreateLogStream",
     ]
 
-    resources = ["${aws_cloudwatch_log_group.es_slow_index_log.arn}"]
+    resources = ["${local.cloudwatch_log_group_arn}"]
 
     principals {
       type        = "Service"
@@ -15,6 +23,8 @@ data "aws_iam_policy_document" "es_slow_index_log" {
 }
 
 resource "aws_cloudwatch_log_resource_policy" "es_slow_index_log" {
+  count = "${var.enable_slow_index_log ? 1 : 0}"
+
   policy_document = "${data.aws_iam_policy_document.es_slow_index_log.json}"
   policy_name     = "${var.slow_index_log_name}"
 }